+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Transport 5.X.X
 
+1. Added support for alternative clock sources during log recording.
+   * [Pull request 340](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/340)
+
 1. Exposed Log and log Playback time information.
    * [Pull request 342](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/342)